fix(schema): map stock_id as Integer to match int4 DB column
Hibernate schema validation failed because stock_id was mapped as Long
while the actual v_allstock_basicprop column is int4.
Updated Germplasm.stockId and all parse sites to Integer/parseInt.

+++ src/main/java/org/irri/snpseek/brapi/controller/CallSetController.java
@@ -178,7 +178,7 @@ public class CallSetController {
             }
             if (germplasmDbId != null && !germplasmDbId.isBlank()) {
                 try {
-                    predicates.add(cb.equal(root.get("stockId"), Long.parseLong(germplasmDbId)));
+                    predicates.add(cb.equal(root.get("stockId"), Integer.parseInt(germplasmDbId)));
                 } catch (NumberFormatException ignored) {
                     predicates.add(cb.disjunction()); // no match
                 }
